#5431 - Allow triggering bulk recommender via remote API
- Split old api controller class up into multiple classes
- Update naming
